[FEAT] #74 - 카드너 보관함 합침, 코드복사 구현
