+++ app/src/main/AndroidManifest.xml
@@ -56,12 +56,10 @@
         <activity
             android:name=".presentation.ui.cardpack.view.CardCreateCompleteActivity"
             android:exported="false" />
-        <activity
-            android:name=".presentation.ui.cardpack.view.TempCardYouStoreActivity"
-            android:exported="true" />
+
         <activity
             android:name=".presentation.ui.login.view.LoginActivity"
-            android:exported="true"/>
+            android:exported="true" />
         <activity
             android:name=".presentation.ui.maincard.view.MainCardActivity"
             android:exported="false" />
